Merge new and old metrics graphs
This commit iterates on the previous two. It deletes the old dashboards
in full favor of the new ones, and modifies json2graphite.rb such that
when old files are injested, they are uploaded with the new, correct
namespace. This means there only needs to be one copy of the dashboards,
and they will work with both old and new files.