Add error metrics to graphs
Prior to this commit no collection errors were displayed in the
dashboards. This commit adds a basic graph of the number of collection
errors seen over time to both the workload and performance dashboards.
This commit is aimed at introduction error metrics into the dashboard
for easier diagnosing of issues.

This commit uses the error count created in
npwalker/pe_metric_curl_cron_jobs#12