Fixed a logic error when handling non API errors
Prior to this commit only API based mbean errors were being included in
the output. With this commit, any error will be parsed and added as a
metric with a hardcoded 1 when it is seen. This allows it to be
graphed in grafana. This commit works for both puppetdb and puppetserver
metric collection.