Updated graphs with links and smaller labels
Prior to this commit the new error metric dashboard did not have any
links and the names of the metrics were very long. This commit adds
links to the other dashboards as well as showing a better name for the
error metrics.

This commit also updates the other dashboards to link back to the error
dashboard