Update docker-compose to use latest image
Prior to this commit each release of this source repository was linked
to a specific reidmv/grafana-puppetserver docker image. This requires
users to update this reposirory every time a new image is pushed to
dockerhub. To reduce the overhead of having to update this repo for
every instance, we should use the lastest image from dockerhub instead.

This commit will use the latest images on docker hub when starting the
program