Cleanup apt cache in docker image
Prior to this commit the apt-cache was updated to install the curl
package, but was never cleaned up. This commit autoremoves packages and
cleans up the apt-cache and /tmp.

This commit makes the docker image 4% smaller in size. While this is not
a large improvement it currently saves 12MB of space for the download
and deployment from an external source.