Add retention argument to view in graphana script
Prior to this commit, the view in graphana script would load
all data privided in a metrics tarball.  Turns out if you have a
full 90 days of metrics and numerous compile masters this can take
a really long time to load ( sometimes it just hangs in the middle
too ).

After this commit, the view in graphana script defaults to only
loading 30 days of data and can be changed by passing a second
argument to the script that is the number of days to load.