Remove json files past retention days if previously extracted
Prior to this commit, if you ran view-in-grafana with 30
retention_days, for example, and then ran it with 15 days you would
end up loading all 30 days of data because we didn't delete any
of the previously extracted json files.

After this commit, we also delete json files past the retention
days.