Extract longest JRuby borrow time
This commit updates the handlers for array-valued metrics to process the
borrowed-instances array returned by Puppet Server's JRuby metrics. The
handler sorts the array and returns the borrow time of the instance that
has been borrowed the longest and reports it as a metric. This allows
JRuby instances that have become "hung" on a request to be identified
as the longest borrow time will keep increasing until the instance is
released.