Add options to the script and update documentation
This commit adds the ability to specify which back end database should
be used in `view-in-grafana.sh`. It enables influxdb to be specified,
but defaults to graphite. It also adds an option to build the graphite
container locally for the purposes of testing dashboard changes. The
documentation has been updated to reflect these changes.