(GH-30) Fix infinite loop
Prior to this commit, if the `nc` command in view-in-grafana was unable
to access the database it would try indefinately. Since we trapped the
ctrl-c it was a pain to stop the process. This commit adds a timeout to
the loop and will fail out if it is unable to connect.